requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.1
python-dotenv==1.0.0
asyncio==3.4.3
pymongo==4.5.0